Rename package to or.http4k.filter to reflect overall project style
